pyproject.toml
# pyproject.toml
# Python project configuration.

[build-system]
requires = ["poetry-core>=1.3.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "buildarr-prowlarr"
version = "0.6.0"
description = "Prowlarr indexer manager plugin for Buildarr"
authors = ["Callum Dickinson <callum.dickinson.nz@gmail.com>"]
license = "GPL-3.0-or-later"
homepage = "https://buildarr.github.io"
repository = "https://github.com/buildarr/buildarr-prowlarr"
documentation = "https://buildarr.github.io/plugins/prowlarr"
keywords = [
    "buildarr",
    "prowlarr",
    "lazylibrarian",
    "lidarr",
    "mylar",
    "radarr",
    "readarr",
    "sonarr",
    "whisparr",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Pydantic",
    "Intended Audience :: End Users/Desktop",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Topic :: System :: Systems Administration",
    "Typing :: Typed",
]
readme = "README.md"
packages = [{include = "buildarr_prowlarr"}]

[tool.poetry.urls]
"Changes" = "https://buildarr.github.io/plugins/prowlarr/release-notes"
"Issue Tracker" = "https://github.com/buildarr/buildarr-prowlarr/issues"

[tool.poetry.plugins."buildarr.plugins"]
"prowlarr" = "buildarr_prowlarr.plugin:ProwlarrPlugin"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
prowlarr-py = ">=1.0.0,<2.0.0"
json5 = ">=0.9.7"
packaging = "*"
buildarr = ">=0.8.a0,<=0.9.0"
buildarr-sonarr = {version = ">=0.6.0", optional = true}
buildarr-radarr = {version = ">=0.3.0", optional = true}

[tool.poetry.group.dev.dependencies]
black = "23.11.0"
mypy = "1.7.0"
types-requests = "2.31.0.10"
mkdocs = "1.5.3"
mkdocstrings = {extras = ["python"], version = "0.23.0"}
ruff = "0.1.5"

[tool.poetry.extras]
sonarr = ["buildarr-sonarr"]
radarr = ["buildarr-radarr"]

[tool.ruff]
fix = true
output-format = "grouped"
target-version = "py38"
line-length = 100
select = [
    "A",
    "B",
    "BLE",
    "E",
    "F",
    "G",
    "I",
    "INP",
    "N",
    "PLC",
    "PLE",
    "PLR",
    "PLW",
    "PTH",
    "RUF",
    "S",
    "T10",
    "T20",
    "W",
    "YTT",
]
extend-select = ["COM812", "COM818", "UP009"]
extend-ignore = [
    "A003",
    "B023",
    "N805",
    "N806",
    "PLR0911",
    "PLR0912",
    "PLR0913",
    "PLR0915",
    "RUF012",
]

[tool.ruff.isort]
lines-between-types = 1
combine-as-imports = true
required-imports = ["from __future__ import annotations"]

[tool.black]
line_length = 100

[tool.mypy]
python_version = "3.8"
pretty = true

[[tool.mypy.overrides]]
module = [
    "json5.*",
    "prowlarr.*",
]
ignore_missing_imports = true
